@@ -22,8 +22,8 @@ const val FB_LOST_FOUND = "lost_n_found"
 const val ADMIN_ID = "Pl4KwImyv9Ro5RQqlT3IqMwXwaH3"
 const val APP_LINK = "https://play.google.com/store/apps/details?id=org.telegram.messenger&hl=en"
 const val TIME_NOTI = "time"
-const val MESS_LINK = "https://drive.google.com/file/d/1C4fAedg5sJ78xzjMvGi46AJA-4TfKx5a/view?usp=sharing"
-const val HOLIDAY_LINK = "https://drive.google.com/file/d/1CCFCfdhEJFHHg3TS_Y3d23haSdOk1YNg/view?usp=sharing"
-const val CALENDER_LINK = "https://drive.google.com/file/d/1t2fTRz18dJcokEO0rFrctqs3kzQQuNDD/view?usp=sharing"
-const val WARDEN_LINK = "https://drive.google.com/file/d/1zV0BDKMvTxhdpTKKK_wTpFRmlQ3FXrVx/view?usp=sharing"
-const val PARENT_PORTAL_LINK = "https://parent.iitism.ac.in/index.php/parent_portal/portal0"
+const val MESS_LINK = "https://md-app-server.herokuapp.com/links/mess"
+const val HOLIDAY_LINK = "https://md-app-server.herokuapp.com/links/holiday"
+const val CALENDER_LINK = "https://md-app-server.herokuapp.com/links/calendar"
+const val WARDEN_LINK = "https://md-app-server.herokuapp.com/links/warden"
+const val PARENT_PORTAL_LINK = "https://md-app-server.herokuapp.com/links/portal"
