@@ -1,6 +1,5 @@
 package com.mailerdaemon.app
 
-import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
@@ -9,16 +8,17 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.DiffUtil
 import com.mailerdaemon.app.R.layout
+import com.mailerdaemon.app.databinding.RvNoticesBinding
+import com.mailerdaemon.app.notices.NoticeAdapter
+import com.mailerdaemon.app.notices.PostModel
 import com.mailerdaemon.app.placement.DiffUtilCallback
 import com.mailerdaemon.app.placement.PlacementActivity
-import com.mailerdaemon.app.placement.PlacementAdapter
-import com.mailerdaemon.app.placement.PlacementModel
 import kotlinx.android.synthetic.main.fragment_placement.*
 import kotlinx.android.synthetic.main.fragment_placement.view.*
 
 class PlacementFragment : Fragment(), PlacementActivity.Companion.ShowData {
-    var data = emptyList<PlacementModel>()
-    private lateinit var adapter: PlacementAdapter
+    var data = emptyList<PostModel>()
+    private lateinit var adapter: NoticeAdapter
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -30,14 +30,10 @@ class PlacementFragment : Fragment(), PlacementActivity.Companion.ShowData {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        adapter = PlacementAdapter(data) { it ->
-            val intent = Intent(activity, PostDetailActivity::class.java)
-            val bundle = Bundle()
-            bundle.putParcelable("post", it)
-            intent.putExtras(bundle)
-            intent.putExtra("type", "Placement")
-            startActivity(intent)
+        val bind: (RvNoticesBinding) -> Unit = {
+            it.noticeDetail.maxLines = 30
         }
+        adapter = NoticeAdapter(data, null, bind)
         (activity as PlacementActivity).getData(this)
         view.refresh.setOnRefreshListener {
                 refresh.visibility = View.GONE
@@ -45,7 +41,7 @@ class PlacementFragment : Fragment(), PlacementActivity.Companion.ShowData {
         }
     }
 
-    override fun showData(list: List<PlacementModel>) {
+    override fun showData(list: List<PostModel>) {
         val n = DiffUtilCallback(data, list)
         val diffResult = DiffUtil.calculateDiff(n)
         data = list
